redirectLogin, Routes, Reset Monthly Spending
redirect Login will be gone in some browser from 24 June.
Safari 16.1+ on macOS
iOS 16.1+
Firefox 109+
Chrome M115+
so I set it to popUpMethode

I set the Routes now properly. Added Some new components to the application.

And finally I fixed a huge problem with the reset button.
I spent over nearly 20 hours just by calculating the value to the income. After that many frustration I discovered that the initialIncome is the value I am trying to get.
So now with little tweaks here and there the Reset button is working as well.